@@ -13,39 +13,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.7",
-    "@angular/common": "5.2.7",
-    "@angular/compiler": "5.2.7",
-    "@angular/core": "5.2.7",
-    "@angular/forms": "5.2.7",
-    "@angular/http": "5.2.7",
-    "@angular/platform-browser": "5.2.7",
-    "@angular/platform-browser-dynamic": "5.2.7",
-    "@angular/router": "5.2.7",
-    "core-js": "2.5.3",
-    "ngx-quill": "2.1.1",
-    "rxjs": "5.5.6",
-    "zone.js": "0.8.20"
+    "@angular/animations": "6.0.1",
+    "@angular/common": "6.0.1",
+    "@angular/compiler": "6.0.1",
+    "@angular/core": "6.0.1",
+    "@angular/forms": "6.0.1",
+    "@angular/http": "6.0.1",
+    "@angular/platform-browser": "6.0.1",
+    "@angular/platform-browser-dynamic": "6.0.1",
+    "@angular/router": "6.0.1",
+    "core-js": "2.5.6",
+    "ngx-quill": "3.1.0",
+    "rxjs": "6.1.0",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.2",
-    "@angular/compiler-cli": "5.2.7",
-    "@angular/language-service": "5.2.7",
-    "@types/jasmine": "2.8.6",
+    "@angular/cli": "6.0.1",
+    "@angular/compiler-cli": "6.0.1",
+    "@angular/language-service": "6.0.1",
+    "@types/jasmine": "2.8.7",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "9.4.6",
-    "codelyzer": "4.2.1",
+    "@types/node": "10.0.8",
+    "codelyzer": "4.3.0",
     "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
+    "karma": "2.0.2",
     "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.1",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.3.0",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.7.2"
+    "karma-cli": " 1.0.1",
+    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "1.1.0",
+    "protractor": "5.3.2",
+    "ts-node": "6.0.3",
+    "tslint": "5.10.0",
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "0.6.1"
   }
 }
