@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { BlogService } from '../../services/blog/blog.service';
 import { PostListItem } from '../../models/post.model';
