@@ -1,8 +1,9 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 import { ActivatedRoute } from '@angular/router';
-import { ISubscription } from 'rxjs/Subscription';
-import 'rxjs/add/operator/switchMap';
+
+import { SubscriptionLike } from 'rxjs';
+import { switchMap } from 'rxjs/operators';
 
 import { BlogService } from '../../services/blog/blog.service';
 import { PostDetail } from '../../models/post.model';
@@ -17,14 +18,16 @@ export class PostComponent implements OnInit, OnDestroy {
 
   public post: PostDetail;
   public content: SafeHtml;
-  private postSubscription: ISubscription;
+  private postSubscription: SubscriptionLike;
 
   constructor(private route: ActivatedRoute, private sanitizer: DomSanitizer, private blogService: BlogService) {}
 
   ngOnInit() {
     this.postSubscription =
       this.route.params
-        .switchMap(params => this.blogService.getPost(params['slug']))
+        .pipe(
+          switchMap(params => this.blogService.getPost(params['slug']))
+        )
         .subscribe(post => {
           this.content = this.sanitizer.bypassSecurityTrustHtml(post.content);
           this.post = post;
